requirements.txt
selenium~=4.8.3
requests~=2.28.2
tabulate~=0.9.0
webdriver-manager~=3.8.5
PyYAML~=6.0